Allow optional parameters for objects in arrays
The JSON → JSchema generator now no longer requires members in arrays to all match the same type. Instead, they must have at least one key in common, which is added to the JSchema. This allows users to generate a JSchema for JSON files that have optional parameters for some objects.